request #23460 Psalm: 4.10.0 -> 4.11.1
Some errors have been fixed, the rest has been integrated into the
baseline.

Changelogs:
https://github.com/vimeo/psalm/releases/tag/4.11.0
https://github.com/vimeo/psalm/releases/tag/4.11.1

+++ src/common/Queue/Redis/BackOffDelayFailedMessage.php
@@ -63,6 +63,6 @@ final class BackOffDelayFailedMessage
     {
         $exponent = min($nb_time_message_has_been_queued - 1, self::MAX_RETRIES_EXPONENT);
 
-        return self::BASE_DELAY_SEC * (2 ** $exponent);
+        return (int) (self::BASE_DELAY_SEC * (2 ** $exponent));
     }
 }
